Change: Increase object/station/roadstop class limit.
The class limit is arbitrary and not stored in game state.

This change prevents all entities in classes after the 255th class from being dumped into the first class.

+++ src/road_cmd.h
@@ -14,7 +14,7 @@
 #include "road_type.h"
 #include "command_type.h"
 
-enum RoadStopClassID : uint8_t;
+enum RoadStopClassID : uint16_t;
 
 void DrawRoadDepotSprite(int x, int y, DiagDirection dir, RoadType rt);
 void UpdateNearestTownForRoadTiles(bool invalidate);
